requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[sqlite]
python-jose
passlib[bcrypt]
passlib[argon2]
python-multipart
pydantic[email]
langchain
langchain-community
langchain-huggingface
faiss-cpu
PyPDF2
celery
redis
google-generativeai
python-dotenv
